rename: mypage page customer 구분 재작명
