+++ src/app/mypages/page.tsx
@@ -1,0 +1,9 @@
+
+export default function MyPage(){
+    return(
+        <>
+            <BannerSelfSeller></BannerSelfSeller>
+            <MyPageContent></MyPageContent>
+        </>
+    );
+}
