+++ src/components/MyPage/customerSelf/BoughtList.tsx
